+++ packages/polythene-core-css/src/layout-styles/common-classes.js
@@ -34,6 +34,13 @@ export default [{
   },
 
   ".pe-rtl": {
-    direction: "rtl"
+    direction: "rtl",
+  },
+
+  // flip directional icon if needed
+  "*[dir=rtl], .pe-rtl ": {
+    " .pe-rtl--flip": {
+      transform: "scaleX(-1)"
+    }
   }
 }];