+++ packages/test-fab/src/tests-generic.js
@@ -1,10 +1,12 @@
 import { FABCSS } from "polythene-css";
 
 const iconAlarm = "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\"><path d=\"M22 5.72l-4.6-3.86-1.29 1.53 4.6 3.86L22 5.72zM7.88 3.39L6.6 1.86 2 5.71l1.29 1.53 4.59-3.85zM12.5 8H11v6l4.75 2.85.75-1.23-4-2.37V8zM12 4c-4.97 0-9 4.03-9 9s4.02 9 9 9c4.97 0 9-4.03 9-9s-4.03-9-9-9zm0 16c-3.87 0-7-3.13-7-7s3.13-7 7-7 7 3.13 7 7-3.13 7-7 7z\"/></svg>";
+const iconBack = "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\"><path d=\"M20 11H7.83l5.59-5.59L12 4l-8 8 8 8 1.41-1.41L7.83 13H20v-2z\"/></svg>";
 
 export default ({ FAB, Icon, renderer: h }) => {
 
   const trustedIconAlarm = h.trust(iconAlarm);
+  const trustedIconBack = h.trust(iconBack);
   
   FABCSS.addStyle(".tests-fab-themed-fab", {
     color_light:                 "#0097A7",
@@ -131,6 +133,26 @@ export default ({ FAB, Icon, renderer: h }) => {
           svg: { content: trustedIconAlarm }
         },
         className: "tests-fab-themed-fab"
+      }
+    },
+
+    {
+      section: "Right-to-left",
+    },
+    {
+      name: "Option: icon (RTL)",
+      interactive: true,
+      component: {
+        view: () => 
+          h("div",
+            { className: "pe-rtl" },
+            h(FAB, {
+              className: "pe-rtl--flip",
+              icon: {
+                svg: { content: trustedIconBack }
+              },
+            })
+          )
       }
     },
   ];